+++ apps/playground/src/app/_layout.tsx
@@ -7,7 +7,6 @@ import { Stack } from 'expo-router/stack'
 import { useEffect } from 'react'
 
 import { AudioFilesProvider } from '../context/AudioFilesProvider'
-import { TranscriberProvider } from '../context/TranscriberContext'
 import { ApplicationContextProvider } from '../context/application-context'
 
 const logger = getLogger('RootLayout')
@@ -24,65 +23,59 @@ export default function RootLayout() {
 
     return (
         <ApplicationContextProvider debugMode>
-            <TranscriberProvider
-                model="Xenova/whisper-tiny"
-                multilingual={false}
-                quantized={false}
+            <AudioRecorderProvider
+                config={{
+                    debug: true,
+                    audioWorkletUrl: `${baseUrl}/audioworklet.js`,
+                    featuresExtratorUrl: `${baseUrl}/audio-features-extractor.js`,
+                }}
             >
-                <AudioRecorderProvider
-                    config={{
-                        debug: true,
-                        audioWorkletUrl: `${baseUrl}/audioworklet.js`,
-                        featuresExtratorUrl: `${baseUrl}/audio-features-extractor.js`,
+                <UIProvider
+                    toastProviderProps={{
+                        overrides: {
+                            snackbarStyle: {
+                                marginBottom: 40,
+                            },
+                        },
                     }}
                 >
-                    <UIProvider
-                        toastProviderProps={{
-                            overrides: {
-                                snackbarStyle: {
-                                    marginBottom: 40,
-                                },
-                            },
-                        }}
-                    >
-                        <AudioFilesProvider>
-                            <Stack
-                                screenOptions={{
-                                    headerBackButtonMenuEnabled: false,
-                                    // headerLeft: ({ label, canGoBack, tintColor }) => {
-                                    //   if (canGoBack) {
-                                    //     return (
-                                    //       <MaterialIcons
-                                    //         name="arrow-back-ios"
-                                    //         size={24}
-                                    //         color={tintColor}
-                                    //         onPress={() => router.back()}
-                                    //         style={{ paddingRight: 10, paddingLeft: 10 }}
-                                    //       />
-                                    //     );
-                                    //   } else {
-                                    //     return (
-                                    //       <MaterialIcons
-                                    //         name="home"
-                                    //         size={24}
-                                    //         color={tintColor}
-                                    //         onPress={() => router.navigate("/")}
-                                    //         style={{ paddingRight: 10, paddingLeft: 10 }}
-                                    //       />
-                                    //     );
-                                    //   }
-                                    // },
-                                }}
-                            >
-                                <Stack.Screen
-                                    name="(tabs)"
-                                    options={{ headerShown: false }}
-                                />
-                            </Stack>
-                        </AudioFilesProvider>
-                    </UIProvider>
-                </AudioRecorderProvider>
-            </TranscriberProvider>
+                    <AudioFilesProvider>
+                        <Stack
+                            screenOptions={{
+                                headerBackButtonMenuEnabled: false,
+                                // headerLeft: ({ label, canGoBack, tintColor }) => {
+                                //   if (canGoBack) {
+                                //     return (
+                                //       <MaterialIcons
+                                //         name="arrow-back-ios"
+                                //         size={24}
+                                //         color={tintColor}
+                                //         onPress={() => router.back()}
+                                //         style={{ paddingRight: 10, paddingLeft: 10 }}
+                                //       />
+                                //     );
+                                //   } else {
+                                //     return (
+                                //       <MaterialIcons
+                                //         name="home"
+                                //         size={24}
+                                //         color={tintColor}
+                                //         onPress={() => router.navigate("/")}
+                                //         style={{ paddingRight: 10, paddingLeft: 10 }}
+                                //       />
+                                //     );
+                                //   }
+                                // },
+                            }}
+                        >
+                            <Stack.Screen
+                                name="(tabs)"
+                                options={{ headerShown: false }}
+                            />
+                        </Stack>
+                    </AudioFilesProvider>
+                </UIProvider>
+            </AudioRecorderProvider>
         </ApplicationContextProvider>
     )
 }
